\name{TrainSel-package}
\alias{TrainSel-package}
\alias{TrainSel-package}
\docType{package}
\title{
\packageTitle{TrainSel}
}
\description{
\packageDescription{TrainSel}

Note: If you have installation issues on a Mac visit:

\url{https://github.com/rmacoslib/r-macos-rtools/}

}
\author{
\packageAuthor{TrainSel}
Maintainer: \packageMaintainer{TrainSel}
}

