\name{UtilityFunctions}
\alias{calculatecrossvalueM1}
\alias{calculatecrossvalueM2}

\title{
Utility Functions
}
\description{
Some utility functions: DOPT MaximinOPT PEVlmOPT CDMEANOPT CDMEANOPTTarget  CDMEANOPTwithEnv CDMEANOPTwithEnvTarget CDMEANOPTwithEnvME funOptProp
}


\author{
Deniz Akdemir, Simon Rio, Julio Isidro Sanchez, Javier Fernandez-Gonzalez
}
