\name{WheatData}
\alias{WheatData}
\alias{Wheat.Y}
\alias{Wheat.M}
\alias{Wheat.K}
\title{Adult plant height (estimated genetic values) and marker dataset for 1182 elite wheat lines}
\description{Containing the phenotypic observations 'Wheat.Y', markers 'Wheat.M' and genetic relationships 'Wheat.K'. The 4670 markers available for these 200 genotypes were pre-porecessed for missingness and minor ellele frequencies, coded numerically as -1, 0, and 1; the relationship genomic relationship matrix was calculated from these markers.}
\source{
		This dataset was obtained from https://triticeaetoolbox.org.
	}
		