\name{UtilityFunctions}
\alias{calculatecrossvalueM1}
\alias{calculatecrossvalueM2}

\title{
Utility Functions
}
\description{
Some utility functions: DOPT MaximinOPT PEVlmOPT CDMEANOPT CDMEANOPTTarget  CDMEANOPTwithEnv CDMEANOPTwithEnvTarget CDMEANOPTwithEnvME funOptProp
}

\details{
Selection by partitioning around medoids.
}
\value{vector of integers that correspond to the selected rows of M.}

\author{
Deniz Akdemir, Simon Rio, Julio Isidro Sanchez
}
